+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 <!-- version list -->
 
+## v0.4.0-dev.2 (2026-09-18)
+
+### Documentation
+
+- Add git installation instructions to docs
+  ([#411](https://github.com/destiny-evidence/data-extraction-evaluation-toolkit/pull/411),
+  [`c14a32d`](https://github.com/destiny-evidence/data-extraction-evaluation-toolkit/commit/c14a32d972ee51c843e2713c5287cd73e48f3329))
+
+### Features
+
+- Don't crash on meeting an eppiattribute that is unkown, but return unkown
+  ([`80de381`](https://github.com/destiny-evidence/data-extraction-evaluation-toolkit/commit/80de381ce9a07ad22795092d234b56cf8c2ca8bb))
+
+
+## v0.4.0-dev.1 (2026-09-18)
+
+
+## v0.3.0-dev.6 (2026-09-18)
+
+### Bug Fixes
+
+- Hoist importlib.metadata import to top-level in version CLI tests
+  ([`aed3c07`](https://github.com/destiny-evidence/data-extraction-evaluation-toolkit/commit/aed3c07abc9af03e64a3b40a0a016a277e8539d3))
+
+### Features
+
+- Add --version and -v flags to CLI
+  ([`0da7ae9`](https://github.com/destiny-evidence/data-extraction-evaluation-toolkit/commit/0da7ae9e0b52abf5acb64be6d5c96e399c4cc144))
+
+
 ## v0.3.0-dev.5 (2026-09-17)
 
 ### Features